a C program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -15,7 +15,7 @@ int main(void)
  printf("Size of a char: %lu byte(s)\n", sizeof(c));
  printf("Size of an int: %lu byte(s)\n", sizeof(i));
  printf("Size of a long int: %lu byte(s)\n", (long)sizeof(i));
- printf("Size of a long long int: %lu byte(s)\n", (long long)sizeof(i));
+ printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(i));
  printf("Size of an float: %lu byte(s)\n", sizeof(f));
  return (0);
 }
